Added a change log function
Added a change log function in the model, and implemented it with the
"add new serial number/system type" function.

I had to edit the splash page to ask for the technician's initials so
that I could test it! I also edited one of your functions to create a
blank record first, then update it with the serial no. and system type.